Exclude scroll method from type generation.
This method's implementation is compatible with the standard HTMLElement
scroll method, but since we can't represent the overloading with a
Closure type, TypeScript doesn't believe that they are compatible.

Since it's already on the HTMLElement base class, the simplest thing to
do here is just remove it from our typings, rather than trying to make
the signatures line up.